Created Filtering for DF File Removed Circular Dependency
